+++ src/view/Notifications.js
@@ -6,6 +6,7 @@ export default class NotificationsApp extends SvelteApplication {
 
   constructor() {
     super({ widgetId: "notifications" });
+    setTimeout(_ => delete ui.windows[this.appId], 1000);
   }
 
   static get defaultOptions() {