+++ src/view/help/HelpDND.svelte
@@ -1,0 +1,7 @@
+<h1>Alpha File Drag and Drop</h1>
+<ul class="ui-text-base-content">
+   <li>You can drop image and video files right on canvas to create tiles.</li>
+   <li>You can drop images onto actor/image thumbs in the tree to change their art.</li>
+   <li>It's possible for the browser too, but entities should belong to an unlocked compendium.</li>
+   <li>Sheet images are also support dropping. [dnd5e native/tidy, pf2e native/dorako]</li>
+</ul>