+++ src/view/settings/PremiumSettings.svelte
@@ -41,5 +41,6 @@
       <li>File tagging</li>
       <li>"Copy image to clipboard" button</li>
       <li>Option to hide Indexing Status widget</li>
+      <li>Changing token art by dropping image file</li>
    </ul>
 </div>